Fix compiler warnings
After adding several additional compiler check, I fixed all of them.
This includes several implicit conversions and shadowed variables.

Several other problems were fixed:

- Removed const data members! See https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/avoid-const-or-ref-data-members.html
- Make sure to follow rule of five/zero.
- Some member-functions should be static instead.
- Delete implict double-overloads from State

In additional, several simple functions/methods were inlined (and some
of them constexprd). Also, the includes were cleaned up.

+++ include/alpaca/AngCorrRejectionSampler.hh
@@ -50,7 +50,7 @@ public:
    * before the algorithm terminates without success and returns \f$\left( 0, 0
    * \right)\f$.
    */
-  AngCorrRejectionSampler(AngularCorrelation &w, const int seed,
+  AngCorrRejectionSampler(AngularCorrelation &w, const unsigned long seed,
                           const unsigned int max_tri = 1000);
 };
